Drop i18n descriptions
Providing descriptions for localization may seem nice, but for now, are redundant and time-consuming. Also, translations will be mostly done in-house (at least for the languages we are capable of translating) because other languages will be introduced way in the future. So for now, we save ourselves a bit of time and attend this issue in the future.